CORDA-2327 Address code review comments
